posting: リファクタ
